pyproject.toml
[project]
name = "agentr-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "jojoo",email = "jawhardjebbi@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard] (>=0.128.0,<0.129.0)",
    "python-jose (>=3.5.0,<4.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "python-multipart (>=0.0.22,<0.0.23)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "sqlmodel (>=0.0.31,<0.0.32)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
